Add ViolationManager for managing getting violations from database.
