+++ ImportSDKDemo/app/src/main/res/layout/activity_view_violations.xml
@@ -14,6 +14,7 @@
         android:layout_height="598dp"
         android:text="TextView"
         android:textColor="#000000"
+        android:gravity="center_horizontal"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.506"
